Changed readme, merged to main generate-visualization file, chart visualization fixes
