+++ toolkit/analysis_and_visualization/classification/README.md
@@ -20,8 +20,8 @@ python generate-visualization.py -h
 ````
 usage: generate-visualization.py  [-h] 
                                   --inference_results INFERENCE_RESULTS
-                                  --image_dir IMAGE_DIR 
                                   --label LABEL
+                                  [--image_dir IMAGE_DIR ]
                                   [--hierarchy HIERARCHY]
                                   [--model_name MODEL_NAME] 
                                   --output_dir OUTPUT_DIR 
@@ -39,11 +39,15 @@ usage: generate-visualization.py  [-h]
   --output_name         output ADAT file name [required]
 ````
 
-* --inference_results : **Result CSV File**
+* --inference_results : **Result CSV File** 
 
 | Image File Name | Ground Truth Label | Output Label 1 | Output Label 2 | Output Label 3 | Output Label 4 | Output Label 5 | Prob 1 | Prob 2 | Prob 3 | Prob 4 | Prob 5 |
-| -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- |
+| --------------- | ------------------ | -------------- | -------------- | -------------- | -------------- | -------------- | ------ | ------ | ------ | ------ | ------ |
 
+
+The  ```"Image File Name"``` field can have full path of image and does not need ```image_dir``` argument to be specified. This also avoids copying of images to ADAT results directory.
+
+ If the images are to be copied to local ADAT results directory, ```image_dir``` argument needs to be specified. 
 
 ## Sample
