--- toolkit/analysis_and_visualization/classification/assets/scripts/resultHistory.js
@@ -1,141 +1,0 @@
-var modelHistories = [{
-    "totalImages": 50000,
-    "totalMismatch": 5461,
-    "passProb": 740203.6123967936,
-    "passCount": 44539,
-    "avgPassProb": 16.619223880122895,
-    "accuracyPer": 89.078,
-    "mismatchPer": 10.922,
-    "avgMismatchProb": 11.437617131331248,
-    "totalFailProb": 62460.82715419994,
-    "netSummaryImages": 50000,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 00:36:45"
-}, {
-    "totalImages": 50000,
-    "totalMismatch": 5461,
-    "passProb": 740203.6123967936,
-    "passCount": 44539,
-    "avgPassProb": 16.619223880122895,
-    "accuracyPer": 89.078,
-    "mismatchPer": 10.922,
-    "avgMismatchProb": 11.437617131331248,
-    "totalFailProb": 62460.82715419994,
-    "netSummaryImages": 50000,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 00:36:50"
-}, {
-    "totalImages": 50000,
-    "totalMismatch": 5461,
-    "passProb": 740203.6123967936,
-    "passCount": 44539,
-    "avgPassProb": 16.619223880122895,
-    "accuracyPer": 89.078,
-    "mismatchPer": 10.922,
-    "avgMismatchProb": 11.437617131331248,
-    "totalFailProb": 62460.82715419994,
-    "netSummaryImages": 50000,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 00:36:55"
-}, {
-    "totalImages": 50000,
-    "totalMismatch": 5461,
-    "passProb": 740203.6123967936,
-    "passCount": 44539,
-    "avgPassProb": 16.619223880122895,
-    "accuracyPer": 89.078,
-    "mismatchPer": 10.922,
-    "avgMismatchProb": 11.437617131331248,
-    "totalFailProb": 62460.82715419994,
-    "netSummaryImages": 50000,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 00:59:56"
-}, {
-    "totalImages": 258,
-    "totalMismatch": 4,
-    "passProb": 187.72820000000007,
-    "passCount": 254,
-    "avgPassProb": 0.7390874015748035,
-    "accuracyPer": 98.44961240310077,
-    "mismatchPer": 1.550387596899225,
-    "avgMismatchProb": 0.300125,
-    "totalFailProb": 1.2005,
-    "netSummaryImages": 258,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 09:47:58"
-}, {
-    "totalImages": 258,
-    "totalMismatch": 4,
-    "passProb": 187.72820000000007,
-    "passCount": 254,
-    "avgPassProb": 0.7390874015748035,
-    "accuracyPer": 98.44961240310077,
-    "mismatchPer": 1.550387596899225,
-    "avgMismatchProb": 0.300125,
-    "totalFailProb": 1.2005,
-    "netSummaryImages": 258,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 10:32:20"
-}, {
-    "totalImages": 258,
-    "totalMismatch": 4,
-    "passProb": 187.72820000000007,
-    "passCount": 254,
-    "avgPassProb": 0.7390874015748035,
-    "accuracyPer": 98.44961240310077,
-    "mismatchPer": 1.550387596899225,
-    "avgMismatchProb": 0.300125,
-    "totalFailProb": 1.2005,
-    "netSummaryImages": 258,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 10:33:21"
-}, {
-    "totalImages": 258,
-    "totalMismatch": 4,
-    "passProb": 187.72820000000007,
-    "passCount": 254,
-    "avgPassProb": 0.7390874015748035,
-    "accuracyPer": 98.44961240310077,
-    "mismatchPer": 1.550387596899225,
-    "avgMismatchProb": 0.300125,
-    "totalFailProb": 1.2005,
-    "netSummaryImages": 258,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 10:33:50"
-}, {
-    "totalImages": 50000,
-    "totalMismatch": 5461,
-    "passProb": 740203.6123967936,
-    "passCount": 44539,
-    "avgPassProb": 16.619223880122895,
-    "accuracyPer": 89.078,
-    "mismatchPer": 10.922,
-    "avgMismatchProb": 11.437617131331248,
-    "totalFailProb": 62460.82715419994,
-    "netSummaryImages": 50000,
-    "totalNoGroundTruth": 0,
-    "modelName": "",
-    "genDate": "2020-10-22 13:01:02"
-}, {
-    "totalImages": 50000,
-    "totalMismatch": 5461,
-    "passProb": 740203.6123967936,
-    "passCount": 44539,
-    "avgPassProb": 16.619223880122895,
-    "accuracyPer": 89.078,
-    "mismatchPer": 10.922,
-    "avgMismatchProb": 11.437617131331248,
-    "totalFailProb": 62460.82715419994,
-    "netSummaryImages": 50000,
-    "totalNoGroundTruth": 0,
-    "modelName": "Generic Model",
-    "genDate": "2020-10-22 13:08:12"
-}]
