+++ toolkit/analysis_and_visualization/classification/assets/styles/style.css
@@ -362,7 +362,8 @@ select {
 
 .column-3 {
     float: left;
-    width: 25%;
+    width: 24.5%;
+    border: 1px solid #ccc;
 }
 
 .marron-header th {
