+++ toolkit/analysis_and_visualization/classification/assets/templates/index.html
@@ -227,9 +227,12 @@
 
     </div>
 
-    <div id="Hierarchy_pass_fail_chart">
+    <div class="center">
+        <div style="display: inline-block;" id="Hierarchy_pass_fail_chart">
 
+        </div>
     </div>
+
 
 
     <!-- Label SUmmary Filter -->
@@ -265,7 +268,7 @@
                     title="Matched 5th"></td>
             <td><input type="text" size="14" id="fl_misclassifiedTop1" onkeyup="filterLabelTable(10,id)"
                     placeholder="Misclassified Top1 Label" title="Misclassified Top1 Label"></td>
-            <td align="center"><button style="background-color:yellow;" onclick="clearLabelFilter()">Clear
+            <td><button style="background-color:yellow;" onclick="clearLabelFilter()">Clear
                     Filter</button></td>
         </tr>
     </table>
